doc: refactor applyForm, use React naming convention, create components folder
